fix: resolve C99 compound literal initialization warnings
* buffer initialization macros now use C99 designated initializers
* silenced spurious clang parentheses warning in implementation

Changed MULLE__BUFFER_FLEXIBLE_DATA and MULLE__BUFFER_INFLEXIBLE_DATA macros to use designated initializers (._storage, ._curr, etc.) instead of positional initialization to avoid C99 warnings about missing braces in compound literals.

+++ src/mulle-buffer.c
@@ -92,7 +92,7 @@ void
    assert( length != INT_MAX);
 
    buffer->_size             = length;
-   buffer->_type             = MULLE_BUFFER_IS_INFLEXIBLE|MULLE_BUFFER_IS_READONLY;
+   buffer->_type             = (unsigned int) MULLE_BUFFER_IS_INFLEXIBLE | (unsigned int) MULLE_BUFFER_IS_READONLY; // windows cast...
    buffer->_allocator        = NULL;
 
    assert( buffer->_sentinel >= buffer->_storage);
